requirements.txt
django
djangorestframework
openai>=1.0.0
PyPDF2>=3.0.0
